Rename dimo to tesla_dimo and restructure tests
Rename the DIMO action package to tesla_dimo (src/actions/dimo -> src/actions/tesla_dimo) and update imports accordingly (e.g. tesla connector now imports from actions.tesla_dimo.interface). Update config/tesla.json5 to use name "tesla_dimo". Reorganize tests by moving many test modules into connector subpackages and adding __init__.py files to those test folders. Remove an obsolete test (tests/actions/move_go2_autonomy/test_idle.py).
