+++ tests/actions/face/connector/test_face_avatar.py
