+++ tests/actions/gps/connector/test_gps_fabric.py
