+++ tests/actions/move_go2_autonomy/connector/test_move_go2_autonomy_unitree_rplidar_sdk.py
