+++ tests/actions/move_serial_arduino/connector/test_move_serial_arduino_serial_arduino.py
