+++ tests/actions/selfie/connector/test_selfie.py
