+++ tests/actions/speak/connector/test_speak_zenoh.py
